Delete the legacy dashboard
