+++ linkerd/admin/src/main/scala/io/buoyant/linkerd/admin/AdminHandler.scala
@@ -26,10 +26,8 @@ object AdminHandler extends HtmlView {
         <div id="navbar" class="collapse navbar-collapse">
           <ul class="nav navbar-nav">
             <li><a href="/delegator">dtab</a></li>
-            <li><a href="/metrics">metrics</a></li>
             <li><a href="/admin/logging">logging</a></li>
             <li><a href="https://linkerd.io/help/">help</a></li>
-            <li><a href="/legacy-dashboard">legacy dashboard</a></li>
           </ul>
 
           <ul class="nav navbar-nav navbar-right">
@@ -56,10 +54,8 @@ object AdminHandler extends HtmlView {
         <div id="navbar" class="collapse navbar-collapse">
           <ul class="nav navbar-nav">
             <li><a href="/delegator">dtab</a></li>
-            <li><a href="/metrics">metrics</a></li>
             <li><a href="/admin/logging">logging</a></li>
             <li><a href="https://linkerd.io/help/">help</a></li>
-            <li><a href="/legacy-dashboard">legacy dashboard</a></li>
           </ul>
           <ul class="nav navbar-nav navbar-right">
             <li>version ${Build.load().version}</li>
