+++ linkerd/admin/src/main/scala/io/buoyant/linkerd/admin/LinkerdAdmin.scala
@@ -28,7 +28,6 @@ class LinkerdAdmin(app: App, linker: Linker, config: LinkerConfig) extends Admin
 
   private[this] def linkerdAdminRoutes: Seq[(String, Service[Request, Response])] = Seq(
     "/" -> new DashboardHandler,
-    "/legacy-dashboard" -> new SummaryHandler(linker),
     "/files/" -> (StaticFilter andThen ResourceHandler.fromDirectoryOrJar(
       baseRequestPath = "/files/",
       baseResourcePath = "io/buoyant/admin",
