+++ linkerd/admin/src/test/scala/io/buoyant/linkerd/admin/LinkerdAdminTest.scala
@@ -56,7 +56,7 @@ class LinkerdAdminTest extends FunSuite with Awaits {
   test("serves twitter-server static files at /admin/files") {
     val _ = new BuoyantMainTest {
       override def main() {
-        val rsp = await(client(Request("/admin/files/css/summary.css")))
+        val rsp = await(client(Request("/admin/files/css/admin.css")))
         assert(rsp.status == Status.Ok)
       }
     }
